+++ skiros2_skill/src/skiros2_skill/core/skill_utils.py
@@ -1,4 +1,5 @@
 from skiros2_common.core.world_element import Element
+from skiros2_world_model.core.world_model_abstract_interface import WmException
 import skiros2_common.core.params as params
 from skiros2_skill.core.processors import Serial, ParallelFf, State
 import skiros2_common.tools.logger as log
@@ -47,14 +48,19 @@ class NodeExecutor():
         self._tracked_params = []
         self._params = params.ParamHandler()
         self._instanciator = instanciator
+        self._last_print = ""
 
     def syncParams(self, params):
         for k, p in params.iteritems():
             vs = p.values
             if p.dataTypeIs(Element):
-                for i, e in enumerate(vs):
-                    if e.getIdNumber() >= 0:
-                        vs[i] = self._wm.get_element(e.id)
+                for i in reversed(range(0, len(vs))):
+                    if vs[i].getIdNumber() >= 0:
+                        try:
+                            vs[i] = self._wm.get_element(vs[i].id)
+                        except WmException:
+                            log.info("[syncParams]", "{} was deleted, removing from parameters".format(vs[i].id))
+                            vs.pop(i)
                 p.values = vs
 
     def trackParam(self, key, prop="", relation="", print_all=False):
@@ -64,7 +70,7 @@ class NodeExecutor():
         self._tracked_params.append((key, prop, relation, print_all))
 
     def _printTracked(self, params, prefix):
-        to_print = prefix
+        to_print = ""
         for key, prop, relation, print_all in self._tracked_params:
             if params.hasParam(key):
                 es = params.getParamValues(key)
@@ -79,8 +85,9 @@ class NodeExecutor():
                         to_print += "{} {} ".format(key, e)
             else:
                 to_print += key + ' not available. '
-        if to_print != prefix:
-            print to_print
+        if to_print and to_print != self._last_print:
+            self._last_print = to_print
+            print prefix + to_print
 
     def setSimulate(self, sim=True):
         self._simulate = sim
@@ -91,7 +98,7 @@ class NodeExecutor():
 
     def mergeParams(self, skill):
         self._params.reset(self._params.merge(skill._params))
-        self._printTracked(self._params, "[mergeParams] ")
+        self._printTracked(self._params, "[{}:mergeParams] ".format(skill.type))
         #print "Merge: {}".format(self._params.printState())
 
     def inferUnvalidParams(self, skill):
@@ -285,32 +292,41 @@ class NodeExecutor():
         """
         skill.specifyParams(self._params)#Re-apply parameters.... Important!
         self.syncParams(skill.params)
-        self._printTracked(skill._params, "[{}Params] ".format(skill._type))
-        if skill.checkHoldCond(self._verbose):
-            if self._verbose:
-                log.info("[ground]", "Hold-conditions fail for skill {}".format(skill.printInfo()))
-            self.processPreempt(skill)
-            skill.checkHoldCond()#This ensure the skill ends printing the failed conditions
-            return skill.state
+        self._printTracked(skill._params, "[{}:SetParams] ".format(skill.type))
         state = self._postExecute(skill)
         if self._verbose:
             log.info("[VisitorExecute]", "{}".format(skill.printState(self._verbose)))
         self.mergeParams(skill)  # Update params
         return state
 
+    def checkHold(self, skill):
+        """
+        @brief Check hold conditions
+        """
+        if skill.checkHoldCond(self._verbose):
+            if self._verbose:
+                log.info("[ground]", "Hold-conditions fail for skill {}".format(skill.printInfo()))
+            self.processPreempt(skill)
+            skill.checkHoldCond()#This ensure the skill ends printing the failed conditions
+            return skill.state
+        else:
+            return State.Running
+
     def processPreempt(self, skill):
         """
         @brief Stop a skill execution. Visit recursively all children
         """
-        skill.specifyParams(self._params)
-        if self._verbose:
-            log.info("[Preempt]", "{}".format(skill.printState(self._verbose)))
-        #Preempt children
-        for c in skill._children:
-            c.visitPreempt(self)
-        #Preempt skill
-        skill.preempt()
-        self.mergeParams(skill)
+        if skill.hasState(State.Running):
+            skill.specifyParams(self._params)
+            self.syncParams(skill.params)
+            if self._verbose:
+                log.info("[Preempt]", "{}".format(skill.printState(self._verbose)))
+            #Preempt children
+            for c in skill._children:
+                c.visitPreempt(self)
+            #Preempt skill
+            skill.preempt()
+            self.mergeParams(skill)
 
 
 class NodeMemorizer: