+++ skiros2_world_model/src/skiros2_world_model/core/world_model_abstract_interface.py
@@ -1,3 +1,6 @@
+class WmException(Exception):
+    pass
+
 class OntologyAbstractInterface:
     def lock(self):
         """ Not implemented in abstract class. """